% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{calcLL}
\alias{calcLL}
\title{Calculate log likelihood for a given heritability}
\usage{
calcLL(hsq, Kva, y, X, reml = TRUE)
}
\arguments{
\item{hsq}{heritability}

\item{Kva}{eigenvalues of K (calculated by \code{\link{eigen_rotation}})}

\item{y}{rotated phenotypes (calculated by \code{\link{eigen_rotation}})}

\item{X}{rotated covariate matrix (calculated by \code{\link{eigen_rotation}})}

\item{reml}{If TRUE, use REML; otherwise use ordinary maximum likelihood.}
}
\value{
The log likelihood value, with the corresponding estimates
of \code{beta} and \code{sigsq} included as attributes.
}
\description{
Calculate the log likelihood for a given value of the heritability, \code{hsq}.
}

