% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{eigen_rotation}
\alias{eigen_rotation}
\title{eigen decomposition + rotation}
\usage{
eigen_rotation(K, y, X = NULL, Kva = NULL, Kve_t = NULL,
  use_cpp = TRUE)
}
\arguments{
\item{K}{Kinship matrix (required if \code{use_cpp=TRUE})}

\item{y}{Phenotypes}

\item{X}{Numeric matrix with covariates. If NULL, use a column
of 1's (for intercept).}

\item{Kva}{Eigenvalues of \code{K} (optional, ignored if \code{use_cpp=TRUE})}

\item{Kve_t}{= transposed eigenvectors of K (optional, ignored if \code{use_cpp=TRUE})}

\item{use_cpp}{= if TRUE, use c++ version of code}
}
\value{
List containing \code{Kva}, \code{Kve_t} and rotated
\code{y} and \code{X}.
}
\description{
Do eigen decomposition of kinship matrix and rotate \code{X} and
\code{y} by that, i.e., pre-multiply by the transpose of the matrix
of eigenvectors. If \code{Kva} and \code{Kve_t} provided, just do
the "rotation".
}
\examples{
data(recla)
e <- eigen_rotation(recla$kinship, recla$pheno[,1], recla$covar)
}
