% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{fitLMM}
\alias{fitLMM}
\title{Fit a linear mixed model}
\usage{
fitLMM(Kva, y, X, reml = TRUE, check_boundary = TRUE, tol = 0.0001,
  use_cpp = TRUE, compute_se = FALSE)
}
\arguments{
\item{Kva}{Eigenvalues of K (calculated by \code{\link{eigen_rotation}})}

\item{y}{Rotated phenotypes (calculated by \code{\link{eigen_rotation}})}

\item{X}{Rotated covariate matrix (calculated by \code{\link{eigen_rotation}})}

\item{reml}{If TRUE, use REML; otherwise use ordinary maximum likelihood.}

\item{check_boundary}{If TRUE, explicitly check log likelihood at 0 and 1.}

\item{tol}{Tolerance for convergence}

\item{use_cpp}{= if TRUE, use c++ version of code}

\item{compute_se}{= if TRUE, return the standard error of the \code{hsq}
estimate using the Fisher Information matrix of the MLE estimate. The
standard error will be in an \code{attr} of  \code{hsq} in the output.
Currently requires \code{use_cpp = FALSE}, and so if \code{compute_se=TRUE}
we take \code{use_cpp=FALSE}.}
}
\value{
List containing estimates of \code{beta}, \code{sigmasq},
\code{hsq}, \code{sigmasq_g}, and \code{sigmasq_e}, as well as the log
likelihood (\code{loglik}). If \code{compute_se=TRUE}, the output also
contains \code{hsq_se}.
}
\description{
Fit a linear mixed model of the form y = Xb + e where e follows a
multivariate normal distribution with mean 0 and variance matrix
\code{sigmasq_g K + sigmasq_e I}, where \code{K} is a known kniship
matrix and \code{I} is the identity matrix.
}
\examples{
data(recla)
e <- eigen_rotation(recla$kinship, recla$pheno[,1], recla$covar)
result <- fitLMM(e$Kva, e$y, e$X)

# also compute SE
wSE <- fitLMM(e$Kva, e$y, e$X, compute_se = TRUE, use_cpp=FALSE)
c(hsq=wSE$hsq, SE=wSE$hsq_se)

}
