% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{fitLMM}
\alias{fitLMM}
\title{Fit a linear mixed model}
\usage{
fitLMM(Kva, y, X, reml = TRUE, tol = 0.0001)
}
\arguments{
\item{Kva}{Eigenvalues of K (calculated by \code{\link{eigen_rotation}})}

\item{y}{Rotated phenotypes (calculated by \code{\link{eigen_rotation}})}

\item{X}{Rotated covariate matrix (calculated by \code{\link{eigen_rotation}})}

\item{reml}{If TRUE, use REML; otherwise use ordinary maximum likelihood.}

\item{tol}{Tolerance for convergence}
}
\value{
List containing estimates of \code{beta}, \code{sigsq},
\code{hsq}, \code{sigsq_g}, and \code{sigsq_e}, as well as the log
likelihood (\code{loglik}).
}
\description{
Fit a linear mixed model of the form y = Xb + e where e follows a
multivariate normal distribution with mean 0 and variance matrix
\code{sigsq_g K + sigsq_e I}, where \code{K} is a known kniship
matrix and \code{I} is the identity matrix.
}

