% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{getMLsoln}
\alias{getMLsoln}
\title{Get MLEs for coefficients and variance}
\usage{
getMLsoln(hsq, Kva, y, X, reml = TRUE)
}
\arguments{
\item{hsq}{heritability}

\item{Kva}{eigenvalues of K (calculated by \code{\link{eigen_rotation}})}

\item{y}{rotated phenotypes (calculated by \code{\link{eigen_rotation}})}

\item{X}{rotated covariate matrix (calculated by \code{\link{eigen_rotation}})}

\item{reml}{If TRUE, use REML; otherwise use ordinary maximum likelihood.}
}
\value{
list containing \code{beta} and \code{sigsq}, with residual
sum of squares and (if \code{reml=TRUE}, \code{log det (X'SX)}) as
attributes.
}
\description{
For a fixed value for \code{hsq}, the heritability, calculate the
corresponding maximum likelihood estimates of \code{beta} and
\code{sigsq}, with the latter being the total variance,
\code{sigsq_g + sigsq_e}.
}
\examples{
data(recla)
e <- eigen_rotation(recla$kinship, recla$pheno[,1], recla$covar)
ml <- getMLsoln(0.5, e$Kva, e$y, e$X)
}

