% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{getMLsoln}
\alias{getMLsoln}
\title{Get MLEs for coefficients and variance}
\usage{
getMLsoln(hsq, Kva, y, X, reml = TRUE, use_cpp = TRUE)
}
\arguments{
\item{hsq}{heritability}

\item{Kva}{eigenvalues of K (calculated by \code{\link[=eigen_rotation]{eigen_rotation()}})}

\item{y}{rotated phenotypes (calculated by \code{\link[=eigen_rotation]{eigen_rotation()}})}

\item{X}{rotated covariate matrix (calculated by \code{\link[=eigen_rotation]{eigen_rotation()}})}

\item{reml}{If TRUE, use REML; otherwise use ordinary maximum likelihood.}

\item{use_cpp}{= if TRUE, use c++ version of code}
}
\value{
list containing \code{beta} and \code{sigmasq}, with residual
sum of squares and (if \code{reml=TRUE}, \code{log det (XSX)}) as
attributes.
}
\description{
For a fixed value for \code{hsq}, the heritability, calculate the
corresponding maximum likelihood estimates of \code{beta} and
\code{sigmasq}, with the latter being the total variance,
\code{sigmasq_g + sigmasq_e}.
}
\examples{
data(recla)
e <- eigen_rotation(recla$kinship, recla$pheno[,1], recla$covar)
ml <- getMLsoln(0.5, e$Kva, e$y, e$X)
}
