% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmlite-package.R
\docType{package}
\name{lmmlite-package}
\alias{lmmlite}
\alias{lmmlite-package}
\title{lmmlite: Simple LMM Fitter for QTL Mapping}
\description{
A port of parts of 'pylmm' for fitting linear mixed models relevant to genome-wide association studies (GWAS) and quantitative trait locus (QTL) mapping.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kbroman.org/lmmlite}
  \item \url{https://github.com/kbroman/lmmlite}
  \item Report bugs at \url{https://github.com/kbroman/lmmlite/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID}) ((closely following the pylmm code by Nick Furlotte))

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID}) ((closely following the pylmm code by Nick Furlotte))
}

}
\keyword{internal}
