% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmlite-package.R
\docType{package}
\name{lmmlite-package}
\alias{lmmlite}
\alias{lmmlite-package}
\title{lmmlite: Simple LMM Fitter for QTL Mapping}
\description{
A port of parts of 'pylmm' for fitting linear mixed models relevant to genome-wide association studies (GWAS) and quantitative trait locus (QTL) mapping.
}
\section{Vignettes}{

\itemize{
\item \href{https://kbroman.org/lmmlite/assets/lmmlite.html}{User guide}
\item \href{https://kbroman.org/lmmlite/assets/compare2pylmm.html}{Results compared}
to \href{https://github.com/nickFurlotte/pylmm}{pylmm} (and also the R package
\href{https://cran.r-project.org/package=regress}{regress}).
\item \href{https://kbroman.org/lmmlite/assets/performance.html}{Performance compared}
to pylmm.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://kbroman.org/lmmlite/}
  \item \url{https://github.com/kbroman/lmmlite}
  \item Report bugs at \url{https://github.com/kbroman/lmmlite/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID}) ((closely following the pylmm code by Nick Furlotte))

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID}) ((closely following the pylmm code by Nick Furlotte))
}

}
\keyword{internal}
