% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recla-data.R
\name{recla}
\alias{recla}
\title{Example kinship, phenotype, and covariate data}
\format{A list with three components, \code{kinship}, \code{pheno},
and \code{covar}, each a matrix.}
\source{
QTL Archive, \url{https://phenome.jax.org/projects/Recla1}
}
\description{
Estimated kinship matrix, phenotypes, and covariates
for diversity outcross mice, taken from Recla et al. (2014) and
Logan et al. (2013)
}
\details{
These are data for a set of diversity outcross mice. The
kinship matrix was calculated from MUGA array data, using the
\href{https://github.com/rqtl/qtl2geno}{qtl2geno} package.
}
\examples{
data(recla)

}
\references{
Recla JM, Robledo RF, Gatti DM, Bult CJ, Churchill GA,
Chesler EJ (2014) Precise genetic mapping and integrative
bioinformatics in Diversity Outbred mice reveals Hydin as a novel
pain gene. Mamm Genome 25:211-222

Logan RW, Robledo RF, Recla JM, Philip VM, Bubier JA, Jay JJ,
Harwood C, Wilcox T, Gatti DM, Bult CJ, Churchill GA, Chesler EJ
(2013) High-precision genetic mapping of behavioral traits in the
diversity outbred mouse population. Genes Brain Behav 12:424-437
}
\keyword{datasets}
